pyproject.toml
[project]
name = "blockchain-project-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.47",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
